add school branch

+++ backend/src/branch/branch.service.ts
@@ -1,6 +1,6 @@
-import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { Model, Types } from 'mongoose';
 import { Branch } from 'src/branch/schemas/branch.schema';
 import { UpdateBranchDto } from 'src/branch/dto/updatebranch.dto';
 import { CreateBranchDto } from 'src/branch/dto/createbranch.dto';
@@ -36,6 +36,21 @@ export class BranchService {
     return newBranch;
   }
 
+  async findBySchoolid(user: User): Promise<Branch[]> {
+    const school_id = user.school_id;
+  
+    if (!school_id) {
+      throw new BadRequestException('School ID is missing');
+    }
+  
+    const objectId = Types.ObjectId.isValid(school_id) ? new Types.ObjectId(school_id) : null;
+    if (!objectId) {
+      throw new BadRequestException('Invalid school ID format');
+    }
+  
+    return await this.branchModel.find({ school_id: objectId });
+  }
+
   async findById(id: string): Promise<Branch> {
     const branch = await this.branchModel.findById(id);
     if (!branch) {
@@ -43,6 +58,7 @@ export class BranchService {
     }
     return branch;
   }
+  
 
   async updateById(id: string, branch: UpdateBranchDto): Promise<Branch> {
     return await this.branchModel.findByIdAndUpdate(id, branch, {
@@ -51,7 +67,24 @@ export class BranchService {
     });
   }
 
+  async deleteBySchoolId(school_id: string): Promise<Branch> {
+    const branchs = await this.branchModel.find({school_id});
+    if (!branchs || branchs.length === 0) {
+      throw new NotFoundException('Branch not found.');
+    }
+    return await this.branchModel.findByIdAndDelete({school_id});
+  }
+
   async deleteById(id: string): Promise<Branch> {
-    return await this.branchModel.findByIdAndDelete(id);
+      const branch = await this.branchModel.findById(id);
+      if (!branch) {
+        throw new NotFoundException('Branch not found.');
+      }
+      await this.schoolModel.findByIdAndUpdate(
+        branch.school_id,
+        { $pull: { branch_id: id } },
+        { new: true }
+      ) 
+      return await this.branchModel.findByIdAndDelete(id);
   }
 }
